requirements.txt
streamlit
unstructured
tiktoken
openai
azure.core.credentials
azure.ai.documentintelligence
PyPDF2